Add files via upload
This is as far as I got with my assignment before I got stuck. I want to submit what I had before I lost the file. And this was all that I had. I did my work in pieces. Therefore, I uploaded a screen shot for each stage of my script. Module1 was just to try out whether the code would loop through the ticker and I had it display in the message box. Module2 was to loop through the ticker and display in a column. Module3 was to loop through the tickers and stop and sum each ticker type and display that in a column. I'm sorry, but this was as far as I got before I got stuck. Thank you for allowing me the extra time to recover the work I had done.